+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "WTW",
+  "name": "wtw",
   "version": "13.0.0",
   "type": "module",
   "scripts": {
@@ -31,7 +31,7 @@
     "@types/react-dom": "18.2.6",
     "@types/react-redux": "7.1.25",
     "@types/testing-library__jest-dom": "5.14.6",
-    "@typescript-eslint/eslint-plugin": "5.60.1",
+    "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "5.57.1",
     "@vitejs/plugin-react": "4.0.1",
     "axios-mock-adapter": "1.21.5",
