+++ src/components/app/app.tsx
@@ -1,0 +1,11 @@
+import MainPage from '../../pages/main-page/main-page';
+import { Film } from '../../types/film';
+
+type AppProps = {
+  films: Film[];
+};
+
+export default function App({films}: AppProps):JSX.Element {
+
+  return <MainPage films={films} />;
+}
