+++ src/components/film-card/film-card.tsx
@@ -1,25 +1,23 @@
-import React from 'react';
-
-type Film = {
-  name: string;
+type FilmCardProps = {
+  filmName: string;
   imageSrc: string;
   imageAlt: string;
 };
 
-export default function FilmCard(props: Film) {
+export default function FilmCard({filmName, imageSrc, imageAlt}: FilmCardProps) {
   return (
     <>
       <div className="small-film-card__image">
         <img
-          src={props.imageSrc}
-          alt={props.imageAlt}
+          src={imageSrc}
+          alt={imageAlt}
           width={280}
           height={175}
         />
       </div>
       <h3 className="small-film-card__title">
         <a className="small-film-card__link" href="film-page.html">
-          {props.name}
+          {filmName}
         </a>
       </h3>
     </>
