+++ src/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './pages/MainPage';
+import App from './pages/main-page/main-page';
 
-import {films} from './api';
+import {films} from './mocks/films';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
