+++ src/pages/film/film.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function Film() {
   return (
     <>
