+++ src/pages/my-list/my-list.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function MyList() {
   return (
     <div className="user-page">
