+++ src/pages/player/player.tsx
