+++ src/pages/sign-in/sign-in.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function SignIn() {
   return (
     <div className="user-page">
