+++ src/types/film.ts
@@ -1,0 +1,7 @@
+export type Film = {
+  id: string;
+  name: string;
+  previewImage: string;
+  previewVideoLink: string;
+  genre: string;
+};
